requirements.txt
flask==3.0.2
python-dotenv
PyJWT
Flask-SQLAlchemy
Flask-Migrate
flask-bcrypt
jsonschema[format]
pytz